requirements.txt
line-bot-sdk
flask
requests==2.18.4